Rename IETFHasher back to DefaultHasher
