+++ ec/src/hashing/tests/suites.rs
@@ -5,7 +5,7 @@ use super::json::SuiteVector;
 use libtest_mimic::{run_tests, Arguments, Outcome, Test};
 
 use crate::hashing::map_to_curve_hasher::HashToField;
-use crate::hashing::tests::IETFHasher;
+use crate::hashing::tests::DefaultHasher;
 use ark_ff::PrimeField;
 use ark_test_curves::bls12_381::Fq;
 use ark_test_curves::bls12_381::Fq2;
@@ -40,12 +40,12 @@ fn suites() {
         let hasher;
         match u.curve.as_str() {
             "BLS12-381 G1" => {
-                hasher =
-                    <IETFHasher<Sha256, 128> as HashToField<Fq>>::new_hash_to_field(dst).unwrap();
+                hasher = <DefaultHasher<Sha256, 128> as HashToField<Fq>>::new_hash_to_field(dst)
+                    .unwrap();
             },
             "BLS12-381 G2" => {
-                hasher =
-                    <IETFHasher<Sha256, 128> as HashToField<Fq2>>::new_hash_to_field(dst).unwrap();
+                hasher = <DefaultHasher<Sha256, 128> as HashToField<Fq2>>::new_hash_to_field(dst)
+                    .unwrap();
             },
             _ => return Outcome::Ignored,
         }
